Simplifying carousel.
