+++ RecycleUI/UI/Views/Carousel/Page/PageViewController.swift
@@ -1,0 +1,24 @@
+//
+//  PageViewController.swift
+//  RecycleUI
+//
+//  Created by Yura on 12/22/22.
+//
+
+import SwiftUI
+import UIKit
+
+struct PageViewController<Page: View>: UIViewControllerRepresentable {
+    var pages: [Page]
+
+    func makeUIViewController(context: Context) -> UIPageViewController {
+        let pageViewController = UIPageViewController(transitionStyle: .scroll, navigationOrientation: .horizontal)
+
+        return pageViewController
+    }
+
+    func updateUIViewController(_ pageViewController: UIPageViewController, context: Context) {
+        pageViewController.setViewControllers(
+            [UIHostingController(rootView: pages[0])], direction: .forward, animated: true)
+    }
+}
