+++ RecycleUI/UI/Views/Shared/Tile.swift
@@ -1,0 +1,20 @@
+//
+//  Tile.swift
+//  RecycleUI
+//
+//  Created by Yura on 12/22/22.
+//
+
+import SwiftUI
+
+struct Tile: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct Tile_Previews: PreviewProvider {
+    static var previews: some View {
+        Tile()
+    }
+}
